EDIT: refactored constraints and validator, added dependency instead of common parent
